+++ docs/changelog.md
@@ -1,6 +1,22 @@
 Change Log
 ===
 
+#v0.13.0
+**2016 July 19**
+
+Test meepworks framework and useful for new developer easily to know the meepshop work flow.
+
+1. Migrate gulp server to webpack dev server
+2. Deprecate test-server (remove next version) and remove gulp server file
+3. Add new webpack test server
+
+#v0.12.0
+**2016 July 19**
+
+1. Migrate old source code
+2. Fix child component error log
+3. remove gulp file, remove jspm file
+
 #v0.11.43
 **2015 Dec 08**
 
@@ -249,7 +265,7 @@ Breaking Changes:
 3. Cache buster is built into app-loader's initial scripts
 
 
-#v0.3.19 
+#v0.3.19
 
 *2015 Mar. 06*